+++ src/assets/styles/style.css
@@ -1,3 +1,15 @@
+/* プロフィール */
+.profile {
+  color: gray;
+  height: 100%;
+  width: 70%;
+  display: flex; 
+  flex-direction: column;
+  
+  
+  
+}
+
 
 
   .about {
@@ -41,4 +53,6 @@
     margin: 40px 0 4px;
     letter-spacing: 0.2em;
   }
+  
+
