--- src/components/Title.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Title = () => {
-    return (
-        <div>
-            <p>二ノ宮ポートフォリオ</p>
-        </div>
-    )
-}
-
-export default Title
